afichage modale

+++ FrontEnd/modal.js
@@ -1,47 +1,54 @@
-//import { errorMessage } from "./login.js";
 const authToken = JSON.parse(localStorage.getItem("token"));
 console.log(authToken)
-const modal = document.querySelector("#modal1");
-//const token = JSON.parse(localStorage.getItem("token")).token;
+let modal = null
 let file = document.querySelector("#file");
 //////open modal///////////
 const openModal = function (e) {
     e.preventDefault();
-    modal.style.display = null;
-    modal.removeAttribute("aria-hidden");
-    modal.setAttribute("aria-modal", "true");
+    modal = document.querySelector(e.target.getAttribute('href'))
+    modal.style.display = null
+    modal.removeAttribute("aria-hidden")
+    modal.setAttribute("aria-modal", "true")
+    modal.addEventListener('click', closeModal)
+    modal.querySelector('.close-modal').addEventListener('click', closeModal)
+    modal.querySelector('.js-modal-stop').addEventListener('click', stopPropagation)
+    modal.querySelector('.btn-delet').addEventListener('click', stopPropagation)
+}
+
+  ///////////close modal//////////
+  const closeModal = function (e) {
+    e.preventDefault()
+    if (modal === null) return
+    e.preventDefault()
+    modal.style.display = 'none'
+    modal.setAttribute("aria-hidden", 'true')
+    modal.removeAttribute("aria-modal")
+    modal.removeEventListener('click', closeModal)
+    modal.querySelector('.close-modal').removeEventListener('click', closeModal)
+    modal.querySelector('.js-modal-stop').removeEventListener('click', stopPropagation)
+    modal = null
 };
-///////////close modal//////////
-const closeModal = function (e) {
-    e.preventDefault();
-    if (modal === null) return;
-    modal.style.display = "none";
-    document.querySelector(".container-modal").style.display = null;
-    document.querySelector(".container-ad-photo-modal").style.display = "none";
-};
-/// on gére la fermeture lors en clique au echap
-window.addEventListener("keydown", function (e) {
-    if (e.key === "Escape" || e.key === "Esc") {
-        closeModal(e);
-    }
-});
-/// on gére la fermeture lors de clique dehors
-modal.addEventListener("click", function (e) {
-    closeModal(e);
-});
 // on éviter la propogation de click d'un enfant à son parent
-modal.children[0].addEventListener("click", function (e) {
-    e.stopPropagation();
+const stopPropagation = function(e) {
+     e.stopPropagation();
+}
+
+///////ecoute boutton ouvrir modale vue ajoujte photo /////////
+document.querySelector(".btn-ad-photo-moda").addEventListener("click", function () {
+    document.querySelector(".container-modal").style.display = "none";
+    document.querySelector(".container-ad-photo-modal").style.display =null;
 });
-///////////////ecoute boutton fermeture modale //////////
-document.querySelectorAll(".js-modal").forEach((a) => {
-    a.addEventListener("click", openModal);
-});
+
 ////////// ecoute boutton retour ////////////
 document.getElementById("icone-return").addEventListener("click", function () {
     document.querySelector(".container-modal").style.display = null;
     document.querySelector(".container-ad-photo-modal").style.display = "none";
 });
+///////ecoute boutton ouvrir modale /////////
+document.querySelectorAll(".js-modal").forEach(a => {
+    a.addEventListener('click', openModal)
+})
+
 //////////affichage photo au modal///////////////
 function affichModal(works) {
     document.querySelector(".galeri-modal").innerHTML = null;
@@ -55,32 +62,34 @@ function affichModal(works) {
         sectionGallery.appendChild(worksElement);
         ///ajoute le boutton de supprime////
         worksElement.innerHTML = ` <div class= "icone-photo-modal">
-            <button class="btn-move"><i class="fa-solid fa-up-down-left-right" style="color: white;"></i></button>
-            <button type="button" class="btn-delet" id="btn-delet-${works[elem].id}" ><i class="fa-solid fa-trash-can" style="color: white;"></i></button></div>`;
-        worksElement.appendChild(imageElement);
+            <button class="btn-move" style ="display :none;"><i class="fa-solid fa-up-down-left-right" style="color: white; "></i></button>
+            <button type="button" class="btn-delet" id="btn-delet-${works[elem].id}" ><i class="fa-solid fa-trash-can" style="color: white;"></i></button></div>`
+            worksElement.appendChild(imageElement);
         worksElement.appendChild(titleElement);
+        // cacher et afficher l'icone move
+        imageElement.addEventListener("mouseover", function(){
+            modal.querySelector(".btn-move").style.display = null;
+        })
         console.log(`${works[elem].id}`)
         //////////////////  fonction supprime projet ///////////////
         function deletPhoto(event) {
             event.preventDefault();
-            fetch(`http://localhost:5678/api/works/${works[elem].id}`, {
-                method: "DELETE",
-                headers: {
-                    "Content-Type": "application/json",
-                    "Authorization": `Bearer ${authToken.token}`,
-                    }
-                })
+            if (confirm("Désirez-vous vraiment supprimer ce projet ?") === true) {
+                fetch(`http://localhost:5678/api/works/${works[elem].id}`, {
+                    method: "DELETE",
+                    headers: {
+                        "Content-Type": "application/json",
+                        "Authorization": `Bearer ${authToken.token}`,
+                        }
+                    })
+                }
         }
         /////////// ecouter boutton supprime photo ///////////
-        document.getElementById(`btn-delet-${works[elem].id}`).addEventListener("click", deletPhoto);
+        const btnDelete = document.getElementById(`btn-delet-${works[elem].id}`)
+        btnDelete.addEventListener("click",deletPhoto)
     }
     console.log(works);
 }
-///////ecoute boutton ouvrir modal et ajoujte photo /////////
-document.querySelector(".btn-ad-photo-moda").addEventListener("click", function () {
-        document.querySelector(".container-modal").style.display = "none";
-        document.querySelector(".container-ad-photo-modal").style.display =null;
-});
 //////////ecoute boutton ajouter photo ///////
 file.addEventListener("change", function (event) {
     const pictureWork = event.target.files;
@@ -105,14 +114,13 @@ function addPhoto(event) {
     const token = JSON.parse(localStorage.getItem("token")).token;
     const titleValue = document.querySelector("#name").value;
     const categoryValue = document.querySelector("#category").value;
-    if (
-        titleValue === "" ||
-        categoryValue === "" ||
-        workPicture === undefined
-    ) {
-        //errorMessage("Veuillez remplir tous les champs.")
+    if (titleValue === "" || categoryValue === "" || workPicture === undefined ) {
         return alert("Veuillez remplir tous les champs.");
     }
+    if (titleValue !== "" && categoryValue !== "" && workPicture !== undefined ) {
+        modal.querySelector("#send-picture").classList.replace("btn-valid","btn-style-vert")
+    }
+    console.log(workPicture)
     const categoriesChoices = ["objets", "appartements", "hotels-restaurants"];
     const categoryChoiceInteger = categoriesChoices.findIndex((choice) => {
         return choice === categoryValue;
@@ -133,29 +141,8 @@ function addPhoto(event) {
         })
         .then((result) => {
             console.log(result);
-        });
+        }); 
 }
 /////////// ecouter boutton valide pour ajouter photo ///////////
  document.querySelector("#send-picture").addEventListener("click", addPhoto);
-// ///////////////////  fonction supprime photo ///////////////
-// function deletPhoto(event) {
-//     //const myDataObject = { userId: 1 };
-//     event.preventDefault();
-//     const id = document.getElementById("id").value;
-//     //console.log(id)
-//     fetch(`http://localhost:5678/api/works/${id}`, {
-//         method: "DELETE",
-//         body: JSON.stringify(myDataObject),
-//         headers: {
-//             //"Authorization": `Bearer ${token}`
-//             "Content-Type": "application/json",
-//         },
-//     })
-//         .then((response) => response.json())
-//         .then((result) => {
-//             console.log(result);
-//         });
-// }
 
-// /////////// ecouter boutton supprime photo ///////////
-// document.querySelector("#btn-delet").addEventListener("click", deletPhoto);
